support a locales: comment directive in extract_strings
We want to support some strings that are only intended to be translated into certain languages. To manage that, add a "//locales:" comment directive, and parse it in the extract_strings command. Anything in a block with this directive will be parsed with a Locales entry in its StringInfo.

This isn't supported in the other commands yet, so it's currently only visible in the .extracted.json file.